requirements.txt
streamlit>=1.35.0
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
numpy>=1.24.0
